requirements.txt
graphics.py==5.0.1.post1
pygame==2.5.2
PyOpenGL==3.1.7
PyOpenGL-accelerate==3.1.7